+++ Coding/Problem-Solving-level-03-in-C++/#03_RowSumInArray.cpp
@@ -2,16 +2,16 @@
 #include <cstdlib>
 using namespace std;
 
-int RandomNumber(int From, int To)
+short RandomNumber(short From, short To)
 {
-  int randNum = rand() % (To - From + 1) + From;
+  short randNum = rand() % (To - From + 1) + From;
   return randNum;
 };
 
 void GetRandom3x3Matrix(int arr[3][3])
 {
-  for (int i = 0; i < 3; i++)
-    for (int j = 0; j < 3; j++)
+  for (short i = 0; i < 3; i++)
+    for (short j = 0; j < 3; j++)
     {
       arr[i][j] = RandomNumber(1, 100);
     }
@@ -20,9 +20,9 @@ void GetRandom3x3Matrix(int arr[3][3])
 void PrintRandom3x3Matrix(int arr[3][3])
 {
   cout << "Your random matrix is: \n";
-  for (int i = 0; i < 3; i++)
+  for (short i = 0; i < 3; i++)
   {
-    for (int j = 0; j < 3; j++)
+    for (short j = 0; j < 3; j++)
     {
       cout << arr[i][j] << " ";
     }
@@ -32,9 +32,9 @@ void PrintRandom3x3Matrix(int arr[3][3])
 
 void GetRowSumInArray(int Matrix[3][3], int ResultArray[])
 {
-  for (int i = 0; i < 3; i++){
-    int Sum = 0;
-    for (int j = 0; j < 3; j++){
+  for (short i = 0; i < 3; i++){
+    short Sum = 0;
+    for (short j = 0; j < 3; j++){
       Sum += Matrix[i][j];
     }
     ResultArray[i] = Sum;
@@ -43,7 +43,7 @@ void GetRowSumInArray(int Matrix[3][3], int ResultArray[])
 
 void PrintArray(int Array[], short ArrayLength)
 {
-  for (int i = 0; i < ArrayLength; i++){
+  for (short i = 0; i < ArrayLength; i++){
     cout << "Row [" << i+1 << "] sum is: ";
     cout << Array[i] << endl;
   }
