+++ Coding/Problem-Solving-level-03-in-C++/#04_ColSum.cpp
@@ -3,16 +3,16 @@
 #include <iomanip>
 using namespace std;
 
-int RandomNumber(int From, int To)
+short RandomNumber(short From, short To)
 {
-  int randNum = rand() % (To - From + 1) + From;
+  short randNum = rand() % (To - From + 1) + From;
   return randNum;
 };
 
 void GetRandom3x3Matrix(int arr[3][3])
 {
-  for (int i = 0; i < 3; i++)
-    for (int j = 0; j < 3; j++)
+  for (short i = 0; i < 3; i++)
+    for (short j = 0; j < 3; j++)
     {
       arr[i][j] = RandomNumber(1, 100);
     }
@@ -21,9 +21,9 @@ void GetRandom3x3Matrix(int arr[3][3])
 void PrintRandom3x3Matrix(int arr[3][3])
 {
   cout << "Your random matrix is: \n";
-  for (int i = 0; i < 3; i++)
+  for (short i = 0; i < 3; i++)
   {
-    for (int j = 0; j < 3; j++)
+    for (short j = 0; j < 3; j++)
     {
       cout << setw(1) << arr[i][j] << " ";
     }
@@ -33,11 +33,11 @@ void PrintRandom3x3Matrix(int arr[3][3])
 
 void PrintEachColSum(int Matrix[3][3])
 {
-  for (int i = 0; i < 3; i++)
+  for (short i = 0; i < 3; i++)
   {
-    int Sum = 0;
+    short Sum = 0;
     cout << "col [" << i + 1 << "] sum is: ";
-    for (int j = 0; j < 3; j++)
+    for (short j = 0; j < 3; j++)
     {
       Sum += Matrix[j][i];
     }
@@ -47,7 +47,7 @@ void PrintEachColSum(int Matrix[3][3])
 
 void PrintArray(int Array[], short ArrayLength)
 {
-  for (int i = 0; i < ArrayLength; i++)
+  for (short i = 0; i < ArrayLength; i++)
   {
     cout << "col [" << i + 1 << "] sum is: ";
     cout << Array[i] << endl;
